fix: test errors caused by deps updates & config changes
CBR-13
