+++ src/lib/api/mutations/RefreshSession.gql.ts
@@ -1,5 +1,5 @@
 import { ApolloClient, gql, type NormalizedCacheObject } from '@apollo/client';
-import type { RefreshInput, RefreshMutation } from './generated/types';
+import type { RefreshInput, RefreshMutation } from '../generated/types';
 
 const REFRESH_SESSION = gql`
 	mutation Refresh($refreshToken: String!) {
@@ -10,7 +10,10 @@ const REFRESH_SESSION = gql`
 	}
 `;
 
-export const refreshSession = async (client: ApolloClient<NormalizedCacheObject>, refreshToken: string) => {
+export const refreshSession = async (
+	client: ApolloClient<NormalizedCacheObject>,
+	refreshToken: string
+) => {
 	const { data, errors } = await client.mutate<RefreshMutation, RefreshInput>({
 		mutation: REFRESH_SESSION,
 		variables: { refreshToken }
