+++ src/lib/components/organisms/Navbar/Navbar.test.ts
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/svelte';
-import { describe, it, expect, vi } from 'vitest';
+import { describe, it, expect, vi, beforeAll, afterAll } from 'vitest';
 import { writable } from 'svelte/store';
 import Navbar from './Navbar.svelte';
 import * as contexts from '$lib/contexts';
